% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccohort}
\alias{ccohort}
\title{Simulated Data for the Comprehensive Cohort}
\format{
A data frame wtih 563 observations and 8 variables:
\describe{
\item{Y}{TBD}
\item{M}{TBD}
\item{R}{TBD}
\item{t}{TBD}
\item{age}{TBD}
\item{womac_bq}{TBD}
\item{expectationb}{TBD}
\item{ChronicPainb}{TBD}
}
}
\usage{
ccohort
}
\description{
Simulated Data for the Comprehensive Cohort
}
\examples{
data(ccohort)
}
\keyword{datasets}
