% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCasual-package.R
\docType{package}
\name{CompCausal-package}
\alias{CompCausal}
\alias{CompCausal-package}
\title{CompCausal: Inferring causal effects in comprehensive cohort studies}
\description{
This package provide R implementations of the estimation methods for
inferring causal effects in comprehensive cohort studies,
in the presence of unmeasured confounding and missing outcomes. The causal
effects include comprehensive cohort causal effects, randomized trial causal effects,
and patient preference causal effect.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UofUEpiBio/Comprehensive_cohort}
  \item \url{https://uofuepibio.github.io/Comprehensive_cohort/}
  \item Report bugs at \url{https://github.com/UofUEpiBio/Comprehensive_cohort/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Xu \email{jennifer.xu@hsc.utah.edu}

Authors:
\itemize{
  \item Daniel Scharfstein \email{daniel.scharfstein@hsc.utah.edu} (\href{https://orcid.org/0000-0001-7482-9653}{ORCID}) [project director]
}

Other contributors:
\itemize{
  \item George Vega Yon \email{g.vegayon@gmail.com} (\href{https://orcid.org/0000-0002-3171-0844}{ORCID}) [contributor]
}

}
\keyword{Causal}
\keyword{Comprehensive}
\keyword{Missing}
\keyword{Unmeasured}
\keyword{at}
\keyword{cohort,}
\keyword{confounding,}
\keyword{effect}
\keyword{random,}
