% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehensivecohort-package.R
\docType{package}
\name{comprehensivecohort-package}
\alias{comprehensivecohort}
\alias{comprehensivecohort-package}
\title{comprehensivecohort: Comprehensive Cohort Models}
\description{
TBD
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UofUEpiBio/Comprehensive_cohort}
  \item Report bugs at \url{https://github.com/UofUEpiBio/Comprehensive_cohort/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Xu \email{jennifer.xu@hsc.utah.edu}

Authors:
\itemize{
  \item Daniel Scharfstein \email{daniel.scharfstein@hsc.utah.edu} (\href{https://orcid.org/0000-0001-7482-9653}{ORCID}) [project director]
}

Other contributors:
\itemize{
  \item George Vega Yon \email{g.vegayon@gmail.com} (\href{https://orcid.org/0000-0002-3171-0844}{ORCID}) [contributor]
}

}
\keyword{internal}
